fix execution button and clean up
